eld-trip-planner/backend/requirements.txt
asgiref==3.8.1
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
sqlparse==0.5.3
tzdata==2025.2
